ai-service/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
spacy>=3.7
scikit-learn>=1.5
transformers>=4.41
pdfminer.six>=20221105
python-docx>=1.1
python-dotenv>=1.0
